# Improve TQueue's performance