Continuation of work from #3343.

### Benchmark results

#### Before changes

```
QueueParallelBenchmark.zioTQueue      thrpt   45  481.958 ± 14.418  ops/s
QueueSequentialBenchmark.zioTQueue    thrpt   45  291.183 ± 3.679   ops/s
QueueBackPressureBenchmark.zioTQueue  thrpt   45  313.379 ± 4.393   ops/s
```

#### After changes

```
QueueParallelBenchmark.zioTQueue      thrpt   45  626.895 ± 6.495  ops/s
QueueSequentialBenchmark.zioTQueue    thrpt   45  333.230 ± 9.095  ops/s
QueueBackPressureBenchmark.zioTQueue  thrpt   45  412.465 ± 8.194  ops/s
```